Add `alert` to help investigate bug
When the app is first installed on iOS, `postMessage` seems not to work. But after we quit the app, it does.

This adds an alert to help us confirm our suspicions that `ReactNativeWebView` and `ReactABI33_0_0NativeWebView` are undefined. And includes the keys that are defined on `window`, in case there's something else we can use.

It will need to be removed before we can release.

This also moves the `(window.ReactNativeWebView || window['ReactABI33_0_0NativeWebView'])` into `executeNativeFunction`, just in case there was some order of operations timing issue. Thought it doesn't seems like that should be a problem.